(PC-18192)[PRO] refactor: router/helpers.js -> hooks/useRedirectLoggedUser
